Conflicts:
	bin/pub/home.html
	emails.csv

+++ src/Controlador/ApagarPerfil.java
@@ -8,19 +8,16 @@ import spark.TemplateViewRoute;
 import java.io.*;
 
 public class ApagarPerfil implements TemplateViewRoute{
-
 	@Override
 	public ModelAndView handle(Request req, Response res) {
-		
-		Usuario usuario = req.session().attribute("usuario_logado");
-		
-		File perfis = new File("Perfis/"+usuario.getMatricula());
-		if(perfis.exists()){
+		Usuario usuario = req.session().attribute("usuario_logado");		
+		File perfis = new File("Perfis/"+usuario.getMatricula()+".csv");
+		File cadastros = new File("Cadastros/"+usuario.getMatricula()+".csv");
+		if(perfis.exists() && cadastros.exists()){
 			perfis.delete();
+			cadastros.delete();
 		}
-		
-		
+		res.redirect("/index.html");
 		return null;
 	}
-
 }
